parser for routine instructions
