+++ src/routine.txt
@@ -1,2 +1,3 @@
 MOVETO [1.25, 0]
 MOVETO [0, 0]
+ROTATETO [45]
